fix: separatea ChoiceSlider from TopicCard
